migrate shell away from events
