+++ harness/determined/common/api/__init__.py
@@ -1,7 +1,12 @@
-from determined.common.api import authentication, errors, metric, request
+from determined.common.api import authentication, errors, metric, request, bindings
 from determined.common.api._session import Session
-from determined.common.api import bindings
-from determined.common.api._util import PageOpts, read_paginated, WARNING_MESSAGE_MAP
+from determined.common.api._util import (
+    PageOpts,
+    read_paginated,
+    WARNING_MESSAGE_MAP,
+    wait_for_ntsc_state,
+    NTSC_Kind,
+)
 from determined.common.api.authentication import Authentication, salt_and_hash
 from determined.common.api.logs import (
     pprint_logs,
